CS300-16Fall/PA1/requirements.txt
Django==1.10.2
image==1.5.5
numpy==1.11.2
Pillow==3.4.2
